Widen the rest of the state upstream keeps as a number
Sweep of every scalar float in the engine's structs against its upstream
counterpart, on the rule that a plain JavaScript property is a double and only
a typed array is float32.  Everything that came back as a property is now a
double: the Dattorro coefficients themselves (0.7f is 0.699999988, not 0.7) and
the reverb wrapper's one-pole pre-lowpass and character coefficients; the
filter's smoothed cutoff and its comparison threshold; the volume envelope's
output gain and the modulation envelope's levels; the voice's gain, pans, LFO
phases, pitch offset and sends; the sample's playback step; the channel's gain,
pan, tuning and custom controllers; and the three smoothing factors.  Delay
buffers, mix buffers and the modulator value cache stay as they are -- those
are float32 and int16 upstream too.

None of it moves a number.  All 103 grid probes and all 31 corpus files render
exactly as before, and the four C tests pass.  It is committed because the
types are what upstream has and because the last four measurable divergences
were each a float narrowing that was invisible until it wasn't, not because it
fixed anything today.

Widening the custom controller array also required widening the reset table it
is memcpy'd from, which the compiler caught as an out-of-bounds read.  That one
was worth the sweep on its own.

What remains on the baseline single-note probe is a -105.75 dBFS transient at
the release boundary, with the comparator reporting no divergence and unity
correlation.  There is no systematic float narrowing left to find.

+++ spessasynth_core/include/spessasynth/synthesizer/dsp/dattorro.h
@@ -31,15 +31,15 @@ SS_DattorroDelayLine *delayLine);
 
 typedef struct {
 	unsigned int preDelay;
-	float preLPF;
-	float inputDiffusion[2];
-	float decay;
-	float decayDiffusion[2];
-	float damping;
-	float excursionRate;
-	float excursionDepth;
-	float gain;
-	float sampleRate;
+	double preLPF;
+	double inputDiffusion[2];
+	double decay;
+	double decayDiffusion[2];
+	double damping;
+	double excursionRate;
+	double excursionDepth;
+	double gain;
+	double sampleRate;
 	double lp[3];
 	double excPhase;
 	unsigned int pDWrite;
